[MOB-35468] Выносит конвертацию конфигураций шаблонов в отдельный класс
