+++ Sources/FigmaGen/Resolvers/RenderParameters/RenderParametersResolver.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+protocol RenderParametersResolver {
+
+    func resolveRenderParameters(
+        templates: [TemplateConfiguration]?,
+        nativeTemplateName: String
+    ) -> [RenderParameters]?
+
+}
